Add context switch

+++ kernel/alloc/liballoc.h
@@ -32,9 +32,9 @@ extern "C" {
 struct	boundary_tag
 {
     unsigned int magic;			//< It's a kind of ...
-    unsigned int size; 			//< Requested size.
-    unsigned int real_size;		//< Actual size.
-    int index;					//< Location in the page table.
+    unsigned long long size; 			//< Requested size.
+    unsigned long long real_size;		//< Actual size.
+    unsigned long long index;					//< Location in the page table.
 
     struct boundary_tag *split_left;	//< Linked-list info for broken pages.
     struct boundary_tag *split_right;	//< The same.
